// src/interfaces/Character.ts
import { CharacterGender } from './CharacterGender';
import { CharacterStatus } from './CharacterStatus';

export interface Character {
  id: number;
  name: string;
  status: CharacterStatus;
  species: string;
  type: string;
  gender: CharacterGender;
  origin: Origin;
  location: Origin;
  image: string;
  episode: string[];
  url: string;
  created: string;
}

interface Origin {
  name: string;
  url: string;
}